LakbayElyu
Lakbay Elyu presents viewers with an immersive experience of an activity, location, or destination through the use of technology. There are tons of different types of virtual tourism offerings, but there's usually a combination of virtual reality, still images, video, audio, narration, interactivity, and other multimedia formats to provide an experience of a destination that a user cannot get through images or websites alone.